Changed Serializer to ModelSerializer in FBV

+++ api/views.py
@@ -3,6 +3,7 @@ from rest_framework.response import Response
 from .models import Brand
 from .serializers import BrandSerializer
 from rest_framework.decorators import api_view
+from rest_framework import  status
 # Create your views here.
 
 
@@ -10,42 +11,54 @@ def home(request):
     return HttpResponse('Django is on track')
 
 
-@api_view(['GET', 'POST', 'PUT'])
+@api_view(['GET', 'POST'])
 def brand_list(request):
+
     if request.method == 'GET':
         obj = Brand.objects.all()
-        print(request.body)
-        print(request.data)
+        print('request.body', request.body)
+        print('request.data', request.data)
+        print('query_params', request.query_params)
         serializer = BrandSerializer(obj, many=True)
         print(serializer.data)
-        return Response(serializer.data)
+        return Response(serializer.data, status=status.HTTP_200_OK)
 
     if request.method == 'POST':
+        print('request.body', request.body)
+        print('request.data', request.data)
+        print('query_params', request.query_params)
         serializer = BrandSerializer(data=request.data)
         if serializer.is_valid():
             serializer.save()
-            return Response(serializer.data)
-        else:
-            print('baddd')
-            return Response(serializer.errors)
+            return Response(serializer.data, status=status.HTTP_201_CREATED)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 
-@api_view(['PUT', 'DELETE'])
+@api_view(['PUT', 'PATCH', 'DELETE'])
 def brand_update(request, pk):
+
     if request.method == 'PUT':
+        obj = Brand.objects.get(id=pk)
+        serializer = BrandSerializer(obj, data=request.data)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data, status=status.HTTP_201_CREATED)
+
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+    if request.method == 'PATCH':
         obj = Brand.objects.get(id=pk)
         serializer = BrandSerializer(obj, data=request.data, partial=True)
         if serializer.is_valid():
             serializer.save()
-            return Response(serializer.data)
-        else:
-            print('bad request')
-            return Response(serializer.errors)
+            return Response(serializer.data, status.HTTP_201_CREATED)
+
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     if request.method == 'DELETE':
         obj = Brand.objects.get(id=pk)
         obj.delete()
-        return Response('deleted successfully')
+        return Response(status.HTTP_200, status=status.HTTP_400_BAD_REQUEST)
